@@ -16,9 +16,9 @@
     "cubic-spline": "^2.0.0",
     "data-forge": "^1.7.3",
     "everpolate": "0.0.3",
-    "expo": "^35.0.0",
-    "expo-device": "~1.0.0",
-    "expo-network": "~1.0.0",
+    "expo": "^36.0.0",
+    "expo-device": "~2.0.0",
+    "expo-network": "~2.0.0",
     "firebase": "^5.11.1",
     "form-data": "^2.5.1",
     "fs": "0.0.1-security",
@@ -30,14 +30,14 @@
     "numeral": "^2.0.6",
     "qs": "^6.8.0",
     "querystring": "^0.2.0",
-    "react": "16.8.0",
-    "react-dom": "^16.12.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
+    "react": "16.9.0",
+    "react-dom": "16.9.0",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-36.0.1.tar.gz",
     "react-native-alert-async": "git+https://github.com/slorber/react-native-alert-async.git",
     "react-native-collapse-view": "^1.0.0",
     "react-native-collapsible": "^1.5.1",
     "react-native-foldview": "^1.2.1",
-    "react-native-gesture-handler": "~1.3.0",
+    "react-native-gesture-handler": "~1.5.0",
     "react-native-gifted-chat": "^0.13.0",
     "react-native-htmlview": "^0.15.0",
     "react-native-keyboard-aware-scroll-view": "^0.8.0",
@@ -47,14 +47,14 @@
     "react-native-pdf": "^6.0.1",
     "react-native-progress": "^3.6.0",
     "react-native-progress-wheel": "^1.0.4",
-    "react-native-reanimated": "~1.2.0",
+    "react-native-reanimated": "~1.4.0",
     "react-native-render-html": "^4.2.0",
-    "react-native-screens": "~1.0.0-alpha.23",
+    "react-native-screens": "2.0.0-alpha.12",
     "react-native-simple-radio-button": "^2.7.3",
     "react-native-slider": "^0.11.0",
     "react-native-snap-carousel": "^3.8.1",
     "react-native-step-indicator": "0.0.11",
-    "react-native-svg": "^9.9.4",
+    "react-native-svg": "9.13.3",
     "react-native-svg-transformer": "^0.14.2",
     "react-native-swipe-list-view": "^2.0.1",
     "react-native-switch-selector": "^2.0.0",
@@ -62,8 +62,8 @@
     "react-native-timeline-flatlist": "^0.6.5",
     "react-native-vector-icons": "^6.6.0",
     "react-native-view-pdf": "^0.10.1",
-    "react-native-web": "^0.12.2",
-    "react-native-webview": "^8.1.2",
+    "react-native-web": "^0.11.7",
+    "react-native-webview": "7.4.3",
     "react-navigation": "^3.12.1",
     "react-navigation-tabs": "^2.7.0",
     "recompose": "^0.30.0",
@@ -73,7 +73,7 @@
     "yarn": "^1.17.3"
   },
   "devDependencies": {
-    "babel-preset-expo": "^5.2.0"
+    "babel-preset-expo": "^8.0.0"
   },
   "private": true,
   "rnpm": {
