@@ -1,5 +1,5 @@
 import React from 'react';
-import { View, Text, StyleSheet, Image, TouchableOpacity, Platform, Modal, Alert} from 'react-native';
+import { View, Text, StyleSheet, Image, TouchableOpacity, Dimensions, Modal, Alert} from 'react-native';
 import { NavigationActions } from 'react-navigation';
 import MaterialCommunityIconsIcon from "react-native-vector-icons/MaterialCommunityIcons";
 import FontAwesome from "react-native-vector-icons/FontAwesome";
@@ -26,7 +26,6 @@ import {
     setColor
  } from '../../../Styles/globalStyle'
 
-import Dimensions from 'Dimensions';
 import Numeral from 'numeral'
 import 'numeral/locales/fr'
 
@@ -130,7 +129,7 @@ class FLTemplateAutocall extends React.Component {
 
   }
 
-  componentWillReceiveProps (props) {
+  UNSAFE_componentWillReceiveProps (props) {
  
     typeof props.isGoodToShow !== 'undefined' ? this.setState({ isGoodToShow : props.isGoodToShow }) : null;
     if (typeof props.nominal !== 'undefined' ) {
