@@ -8,7 +8,7 @@ import {
 
 export default class SwipeGesture extends Component {
 
-  componentWillMount = () => {
+  UNSAFE_componentWillMount = () => {
     this.PanResponder = PanResponder.create({
       onStartShouldSetPanResponder: (evt, gestureState) => true,
       onPanResponderRelease: (evt, gestureState) => {
