@@ -1,5 +1,5 @@
 import React from 'react'
-import { InteractionManager, YellowBox, ImageBackground, StyleSheet, Platform } from 'react-native'
+import { InteractionManager, YellowBox, ImageBackground, StyleSheet, Platform, Dimensions} from 'react-native'
 import * as Font from 'expo-font';
 
 import getTheme from './native-base-theme/components'
@@ -8,7 +8,7 @@ import material from './native-base-theme/variables/commonColor'
 import NotificationProvider from './Session/NotificationProvider'
 import Firebase, { FirebaseContext } from './Database';
 
-import Dimensions from 'Dimensions';
+
 
 import Application from './Components/Login/Application'
 import Navigation from './Navigation/Navigation'
