@@ -1,5 +1,5 @@
 import React from 'react';
-import { View, Text, StyleSheet, Image, TouchableOpacity, Platform, Modal, Animated} from 'react-native';
+import { View, Text, StyleSheet, Image, TouchableOpacity, Dimensions, Modal, Animated} from 'react-native';
 import { NavigationActions } from 'react-navigation';
 import MaterialCommunityIconsIcon from "react-native-vector-icons/MaterialCommunityIcons";
 import FontAwesome from "react-native-vector-icons/FontAwesome";
@@ -26,7 +26,7 @@ import {
     setColor
  } from '../../../Styles/globalStyle'
 
-import Dimensions from 'Dimensions';
+
 import Numeral from 'numeral'
 import 'numeral/locales/fr'
 
